refactor into router
